We don't want to force config settings on someone.
There should be defaults for all the various config settings.. This line attempted to set them, but did it in a way that would have overwritten any config settings someone else might have put in.  We will address the need for defaults later and do it in a non-destructive way.